+++ lib/text.ts
@@ -146,11 +146,7 @@ export class AstText extends AstNode {
 	lint(start = this.getAbsoluteIndex(), errorRegex?: RegExp): LintError[] {
 		const {data, parentNode, nextSibling, previousSibling} = this;
 		if (!parentNode) {
-			/* NOT FOR BROWSER */
-
 			throw new Error('无法对孤立文本节点进行语法分析！');
-
-			/* NOT FOR BROWSER END */
 		}
 		const {type, name, parentNode: grandparent} = parentNode;
 		let isHtmlAttrVal = false;